(ceeac) remove obsolete comments
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@8951 8aca7d54-2c30-db11-9de9-000461428c89

+++ descriptor/building_desc.h
@@ -20,13 +20,8 @@ class tool_t;
 class karte_t;
 class checksum_t;
 
-
-/*
- *  Autor:
- *      Volker Meyer
- *
- *  Description:
- *      Data for one tile of a potentially multi-tile building.
+/**
+ * Data for one tile of a potentially multi-tile building.
  *
  *  Child nodes:
  *   0   Imagelist2D season 0 back
@@ -106,12 +101,8 @@ public:
 	uint8 get_layout() const;
 };
 
-/*
- *  Autor:
- *      Volker Meyer
- *
- *  Description:
- *       Data for one building, consists of potentially more than one tile.
+/**
+ * Data for one building, consists of potentially more than one tile.
  *
  *  Child nodes:
  *	0   Name
@@ -194,12 +185,13 @@ class building_desc_t : public obj_desc_timelined_t {
 		// Signal groups for signal boxes
 	koord  size;
 	flag_t flags;
-	uint16 level;			// or passengers;
-	uint8  layouts;			// 1 2, 4, 8  or 16
-	uint16 enables;			// if it is a stop, what is enabled; if it is a signal box, the signal group that can be linked to this box.
-	uint8  distribution_weight;			// Hajo:chance to build, special buildings, only other is weight factor
+	uint16 level;               // or passengers;
+	uint8  layouts;             // 1 2, 4, 8  or 16
+	uint8  enables;             // if it is a stop, what is enabled ...
+	uint8  distribution_weight; // chance to build, special buildings, only other is weight factor
 
-	/** @author: jamespetts.
+
+	/**
 	 * Additional fields for separate capacity/maintenance
 	 * If these are not specified in the .dat file, they are set to
 	 * PRICE_MAGIC then calculated from the "level" in the old way.
@@ -307,13 +299,11 @@ public:
 
 	/**
 	* the level is used in many places: for price, for capacity, ...
-	* @author Hj. Malthaner
 	*/
 	uint16 get_level() const { return level; }
 
 	/**
 	 * Mail generation level
-	 * @author Hj. Malthaner
 	 */
 	uint16 get_mail_level() const;
 
@@ -348,7 +338,6 @@ public:
 
 	/**
 	* Skin: cursor (index 0) and icon (index 1)
-	* @author Hj. Malthaner
 	*/
 	const skin_desc_t * get_cursor() const {
 		return flags & FLAG_HAS_CURSOR ? get_child<skin_desc_t>(2 + size.x * size.y * layouts) : 0;
@@ -370,14 +359,12 @@ public:
 	uint16 get_allowed_region_bits() const { return allowed_regions; }
 
 	/**
-	* @return station flags (only used for station buildings, oil rigs and traction types in depots)
-	* @author prissi
+	* @return station flags (only used for station buildings and oil rigs)
 	*/
 	uint16 get_enabled() const { return enables; }
 
 	/**
 	* @return time for doing one step
-	* @author prissi
 	*/
 	uint16 get_animation_time() const { return animation_time; }
